requirements.txt
beautifulsoup4==4.10.0
requests==2.28.1
nltk==3.8.1
GitPython==3.1.31
python-dateutil==2.8.2